% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoMM.R
\name{CoMM}
\alias{CoMM}
\title{CoMM.}
\usage{
CoMM(
  Model = "Individual",
  file1 = NULL,
  file2 = NULL,
  file3 = NULL,
  file4 = NULL,
  file5 = NULL,
  file6 = NULL,
  Cov1 = NULL,
  Cov2 = NULL,
  R1 = NULL,
  R2 = NULL,
  whCol = 1,
  bw = 5e+05,
  lam = 0.95,
  paral = FALSE,
  coreNum = 1
)
}
\arguments{
\item{Model}{an character specifying which model to run. "Individual" indicates the model using individual eQTL and GWAS data.
"Summary" indicates the model using individual eQTL and summary GWAS data. "Trans" indicates the model using both eQTL and GWAS data
with different reference panel for eQTL and GWAS data. "Shared" indicates the model using both eQTL and GWAS data with the shared reference panel.}

\item{file1}{is the name of transcriptome genotype file.}

\item{file2}{is the name of transcriptome gene expression file.}

\item{file3}{is the name of GWAS plink file.}

\item{file4}{is the name of the summary-level transcriptome data.}

\item{file5}{is the name of complementary summary-level transcriptome data.}

\item{file6}{is the name of summary-level GWAS data.}

\item{Cov1}{is the name of corvariates for transcriptome.}

\item{Cov2}{is the name of corvariates for GWAS.}

\item{R1}{is the name of reference panel for transcriptome.}

\item{R2}{is the name of reference panel for GWAS.}

\item{whCol}{is an integer specifying which column of phenotype. The default is 1.}

\item{bw}{is an integer specifying the number of downstream and upstream SNPs that are considered as cis-SNP within a gene. The default is 500000.}

\item{lam}{is a positive number specifying the shrinkage intensity for reference panel. The default is 0.95.}

\item{paral}{is a logical value specifying whether to run the model in parallel. The default is FALSE.}

\item{coreNum}{is an integer specifying the number of cores for parallel. The default is 1.}
}
\value{
a list, We briefly explain the output of the CoMM.
}
\description{
CoMM implement the model CoMM, CoMM-S2 and CoMM-S4.
}
\details{
CoMM implement the model CoMM, CoMM-S2 and CoMM-S4.
}
