% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CoMM_covar_pxem}
\alias{CoMM_covar_pxem}
\title{CoMM}
\usage{
CoMM_covar_pxem(y, z, x1, x2, w1, w2, sigma2beta, sigma2y, beta0, constr,
  epsStopLogLik = 1e-05, maxIter = 1000L)
}
\arguments{
\item{y}{a vector for the expression of a gene.}

\item{z}{a vector for the phenotype of GWAS.}

\item{w1}{a matrix of coveriates for the expression of a gene.}

\item{w2}{a matrix of coveriates for the GWAS.}

\item{sigma2beta}{a initial value of sigma2beta.}

\item{sigma2y}{a initial value of sigma2y.}

\item{beta0}{a initial value of beta0.}

\item{constr}{indicator value for constraint. When constr = 1, alpha = 0, otherwise, alpha is free to vary.}

\item{epsStopLogLik}{convergence criteria. default is 1e-5.}

\item{maxIter}{maximum iteraion number. default is 1000.}

\item{X1}{a standardized genotype matrix for eQTL data.}

\item{X2}{a standardized genotype matrix for GWAS data.}
}
\value{
List of model parameters
}
\description{
fit CoMM for a single gene
}
