% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CoMM_S2_paral_testing}
\alias{CoMM_S2_paral_testing}
\title{CoMM}
\usage{
CoMM_S2_paral_testing(stringname1, stringname2, stringname3, stringname4,
  stringname5, bw, lam, coreNum)
}
\arguments{
\item{stringname1}{prefix for eQTL genotype file with plink format (bim, bed).}

\item{stringname2}{summary data for GWAS.}

\item{stringname3}{prefix for reference panel file with plink forma (bim,bed).}

\item{stringname4}{expression file for eQTL data.}

\item{stringname5}{covariates file for eQTL data, e.g. top 10 PCs.}

\item{bw}{the number of downstream and upstream SNPs that are considered as cis-SNP within a gene.}

\item{lam}{shirnkage intensity for reference panel}

\item{coreNum}{the number of cores in parallel running}

}

\value{
List of model parameters
}
\description{
CoMM_S2 to dissecting genetic contributions to complex traits by leveraging regulatory information using GWAS summary data.
}
\details{
\code{CoMM_S2_paral_testing} fits the CoMM_S2 model. It requires to provide plink binary eQTL genotype file (bim, bed)
the GWAS summary data, gene expression file for eQTL, reference panel data.

}
\examples{

file1 = "1000G.EUR.QC.1";
file2 = "NFBC_beta_se_TG";
file3 = "1000G_chr_all"
file4 = "Geuvadis_gene_expression_qn.txt";
file5 = "";

bw = 500000;
lam = 0.95;
coreNum = 24;

fm = CoMM_S2_paral_testing(file1,file2,file3, file4,file5, bw, lam, coreNum);

}
\author{
Jin Liu, \email{jin.liu@duke-nus.edu.sg}
}
