% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rl_skat.R
\name{rl_skat}
\alias{rl_skat}
\title{CoMM}
\usage{
rl_skat(Z, y, X = NULL, intercept = T, method = NULL)
}
\arguments{
\item{Z}{a standardized genotype matrix for GWAS data.}

\item{y}{a vector for the phenotype of GWAS.}

\item{Z}{a matrix of coveriates.}
}
\value{
List of model parameters: test statistics (scores) and p-value (pval)
}
\description{
fit SKAT using RL-SKAT: An Exact and Efficient Score Test for Heritability and Set Tests.
}
