% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CoMM_covar_pxem}
\alias{CoMM_covar_pxem}
\title{fm = CoMM_covar_pxem(y,z,x1p,x2p,w1,w2);}
\usage{
CoMM_covar_pxem(y, z, x1, x2, w1, w2, constr = 0L, epsStopLogLik = 1e-05,
  maxIter = 1000L, pxem_indicator = 1L)
}
\description{
fm = CoMM_covar_pxem(y,z,x1p,x2p,w1,w2);
}
\details{
\code{CoMM} fits the CoMM model. It requires to provide plink binary eQTL genotype file (bim, bed)
the GWAS plink binary file (bim, bed, fam), gene expression file for eQTL.
}
